feat (frontend): restructure agent and workflow routes, introduce Agents page
- Removed the ApiKeys page in favor of integrating API key management within the Agents page.
- Replaced Home component with Agents page and updated routing to reflect new structure.
- Introduced AGENTS_BASE_ROUTE and WORKFLOWS_BASE_ROUTE constants for cleaner route management.
- Updated links and routes in various components to align with the new structure.
