+++ frontend/src/config.ts
@@ -1,1 +1,4 @@
 export const API_URL = import.meta.env.VITE_API_URL;
+
+export const AGENTS_BASE_ROUTE = "managed-agents"
+export const WORKFLOWS_BASE_ROUTE = "workflow"
